pyproject.toml
[project]
name = "ai-agent-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "crewai>=1.6.1",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "trafilatura>=2.0.0",
    "uvicorn[standard]>=0.42.0",
]
